+++ source/EngineKernels/DebugKernels.cuh
@@ -7,5 +7,5 @@
 
 
 __global__ void DEBUG_checkAngles(SimulationData data);
-__global__ void DEBUG_checkCellsAndParticles(SimulationData data, float* sumEnergy, int location);
+__global__ void DEBUG_checkCellsAndParticles(SimulationData data, double* sumEnergy, int location);
 __global__ void DEBUG_kernel(SimulationData data, int location);
